+++ pkg/shelf/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.5.2
+
+* Add a `Cascade` helper that runs handlers in sequence until one returns a
+  response that's neither a 404 nor a 405.
+
 ## 0.5.1+1
 
 * Capture all asynchronous errors thrown by handlers if they would otherwise be